springer tutorial in differential equations attempt
Mabye it would help if mathematics was taught in a more programmatic way, because it would allow the user to get an algorithmic perspective on an otherwise simple seeming equation. This was the case when I attempted to get java to return the derivatives of a logarithmic function. In cases like these, functions like matlab and r are better for use for things like that. In running through this tutorials, I fixed many errors, and even went one step further, and I wrote a script that converts mathematical functions into sound.